assignment-python/backend/requirements.txt
Django==2.1.4
pandas==0.24.2
djangorestframework==3.9.1
django-cors-headers==2.5.2